move hooks and funct to utils
