+++ src/components/Catalogues/CataloguePage.js
@@ -1,5 +1,5 @@
 // @flow
-import React, { useEffect, useState, useRef } from "react";
+import React from "react";
 import Waypoint from "react-waypoint";
 import { Helmet } from "react-helmet";
 import {
@@ -24,6 +24,12 @@ import queryString from "query-string";
 import BackButton from "components/BackButton";
 import type { FilterAnyType } from "types/filters";
 import type { Manga } from "@tachiweb/api-client";
+import {
+  jsonToURI,
+  stringToURI,
+  useSearchQueryFromQueryParam,
+  useFiltersFromQueryParam
+} from "./utils";
 
 const useStyles = makeStyles({
   filterButton: {
@@ -138,7 +144,7 @@ const CataloguePage = () => {
         <LocalStateSearchBar
           value={searchQuery}
           onSubmit={handleSearchSubmit}
-          textFieldProps={{ label: "Search for manga" }}
+          textFieldProps={{ label: "Search all catalogues" }}
         />
         <DynamicSourceFilters
           filters={filters}
@@ -177,69 +183,5 @@ const CataloguePage = () => {
     </>
   );
 };
-
-function useSearchQueryFromQueryParam(searchQueryParam: ?string) {
-  const [searchQuery, setSearchQuery] = useState<string>("");
-
-  useEffect(() => {
-    const newSearchQuery = searchQueryParam
-      ? uriToString(searchQueryParam)
-      : "";
-    setSearchQuery(newSearchQuery);
-  }, [searchQueryParam]);
-
-  return searchQuery;
-}
-
-function useFiltersFromQueryParam(
-  initialFilters: FilterAnyType[],
-  filtersQueryParam: ?string
-) {
-  const [filters, setFilters] = useState<FilterAnyType[]>([]);
-
-  // Save a copy of initial filters into state when viewing this page with no filters in the search query
-  const alreadySavedFiltersRef = useRef(false);
-  useEffect(() => {
-    if (initialFilters == null) return;
-    if (filtersQueryParam != null) return;
-    if (alreadySavedFiltersRef.current) return;
-
-    setFilters(initialFilters);
-    alreadySavedFiltersRef.current = true;
-  }, [initialFilters, filtersQueryParam]);
-
-  useEffect(() => {
-    const newFilters = filtersQueryParam
-      ? (uriToJSON(filtersQueryParam): FilterAnyType[])
-      : [];
-    setFilters(newFilters);
-  }, [filtersQueryParam]);
-
-  return filters;
-}
-
-// URL search query can't support nested objects like filters. So encode it when putting it into the URL and vice versa.
-// https://stackoverflow.com/questions/9909620/convert-javascript-object-into-uri-encoded-string
-/**
- * Seems to support arrays in addition to objects
- */
-function jsonToURI(json: Object): string {
-  return encodeURIComponent(JSON.stringify(json));
-}
-
-/**
- * Seems to support arrays in addition to objects
- */
-function uriToJSON(urijson: string): Object {
-  return JSON.parse(decodeURIComponent(urijson));
-}
-
-function stringToURI(string: string): string {
-  return encodeURIComponent(string);
-}
-
-function uriToString(uriString: string): string {
-  return decodeURIComponent(uriString);
-}
 
 export default CataloguePage;
